# artifacts/api-server/src/routes/pitches.ts
import { Router, type IRouter } from "express";
import { and, desc, eq, sql } from "drizzle-orm";
import { db, pitchesTable, pitchBackersTable } from "@workspace/db";
import { currentUserId } from "../lib/currentUser";
import { createNotification } from "../lib/notify";

const router: IRouter = Router();

router.get("/pitches", async (req, res): Promise<void> => {
  const meId = currentUserId(req);
  const all = await db
    .select()
    .from(pitchesTable)
    .orderBy(desc(pitchesTable.trending), desc(pitchesTable.createdAt));
  const mine = await db
    .select()
    .from(pitchBackersTable)
    .where(eq(pitchBackersTable.userId, meId));
  const set = new Set(mine.map((m) => m.pitchId));
  res.json(
    all.map((p) => ({
      ...p,
      coverKey: p.coverKey ?? null,
      backed: set.has(p.id),
    })),
  );
});

router.post("/pitches/:id/back", async (req, res): Promise<void> => {
  const meId = currentUserId(req);
  const id = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const amount = parseInt(String(req.body?.amount ?? "0"), 10);
  if (!id || !Number.isFinite(amount) || amount < 0) {
    res.status(400).json({ error: "Invalid input" });
    return;
  }

  const [pitch] = await db
    .select()
    .from(pitchesTable)
    .where(eq(pitchesTable.id, id));
  if (!pitch) {
    res.status(404).json({ error: "Not found" });
    return;
  }

  const existing = await db
    .select()
    .from(pitchBackersTable)
    .where(
      and(
        eq(pitchBackersTable.pitchId, id),
        eq(pitchBackersTable.userId, meId),
      ),
    );
  if (existing.length === 0) {
    await db
      .insert(pitchBackersTable)
      .values({ pitchId: id, userId: meId });
    await db
      .update(pitchesTable)
      .set({
        backersCount: sql`${pitchesTable.backersCount} + 1`,
        raised: sql`${pitchesTable.raised} + ${amount}`,
      })
      .where(eq(pitchesTable.id, id));
    await createNotification({
      userId: pitch.founderId,
      type: "pitch_backed",
      actorId: meId,
      pitchId: id,
      amount,
      message:
        amount > 0
          ? `expressed interest in ${pitch.title} ($${amount.toLocaleString()})`
          : `expressed interest in ${pitch.title}`,
    });
  }

  const [updated] = await db
    .select()
    .from(pitchesTable)
    .where(eq(pitchesTable.id, id));
  res.json({ ...updated, coverKey: updated.coverKey ?? null, backed: true });
});

export default router;