# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HumanVerse API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface User {
  id: string;
  handle: string;
  name: string;
  title: string;
  company: string;
  city: string;
  country: string;
  avatarKey: string;
  verified: boolean;
  followersCount: number;
  bio: string;
  following: boolean;
}

export interface FollowState {
  userId: string;
  following: boolean;
  followersCount: number;
}

export interface Post {
  id: string;
  authorId: string;
  text: string;
  /** @nullable */
  imageKey?: string | null;
  sponsored: boolean;
  /** @nullable */
  sponsorLabel?: string | null;
  category: string;
  likesCount: number;
  retweetsCount: number;
  commentsCount: number;
  tipsTotal: number;
  createdAt: string;
  liked: boolean;
  retweeted: boolean;
}

export interface CreatePostBody {
  text: string;
  /** @nullable */
  imageKey?: string | null;
}

export interface TipBody {
  amount: number;
}

export interface Circle {
  id: string;
  name: string;
  about: string;
  category: string;
  color: string;
  /** @nullable */
  coverKey?: string | null;
  paid: boolean;
  price: number;
  membersCount: number;
  activeNow: number;
  founderIds: string[];
  joined: boolean;
}

export interface Pitch {
  id: string;
  founderId: string;
  title: string;
  stage: string;
  industry: string;
  raising: number;
  raised: number;
  city: string;
  summary: string;
  /** @nullable */
  coverKey?: string | null;
  backersCount: number;
  trending: boolean;
  backed: boolean;
}

export interface BackPitchBody {
  amount: number;
}

export interface Marker {
  id: string;
  type: string;
  label: string;
  city: string;
  x: number;
  y: number;
  meta: string;
  /** @nullable */
  refId?: string | null;
}

export interface Notification {
  id: number;
  userId: string;
  type: string;
  /** @nullable */
  actorId?: string | null;
  /** @nullable */
  postId?: string | null;
  /** @nullable */
  circleId?: string | null;
  /** @nullable */
  pitchId?: string | null;
  /** @nullable */
  amount?: number | null;
  message: string;
  read: boolean;
  createdAt: string;
}

export type ListPostsParams = {
  feed?: ListPostsFeed;
};

export type ListPostsFeed = (typeof ListPostsFeed)[keyof typeof ListPostsFeed];

export const ListPostsFeed = {
  foryou: "foryou",
  following: "following",
  investors: "investors",
  hiring: "hiring",
} as const;