# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HumanVerse API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.string(),
  handle: zod.string(),
  name: zod.string(),
  title: zod.string(),
  company: zod.string(),
  city: zod.string(),
  country: zod.string(),
  avatarKey: zod.string(),
  verified: zod.boolean(),
  followersCount: zod.number(),
  bio: zod.string(),
  following: zod.boolean(),
});

/**
 * @summary List all users
 */
export const ListUsersResponseItem = zod.object({
  id: zod.string(),
  handle: zod.string(),
  name: zod.string(),
  title: zod.string(),
  company: zod.string(),
  city: zod.string(),
  country: zod.string(),
  avatarKey: zod.string(),
  verified: zod.boolean(),
  followersCount: zod.number(),
  bio: zod.string(),
  following: zod.boolean(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Toggle follow on a user
 */
export const ToggleFollowParams = zod.object({
  id: zod.coerce.string(),
});

export const ToggleFollowResponse = zod.object({
  userId: zod.string(),
  following: zod.boolean(),
  followersCount: zod.number(),
});

/**
 * @summary List feed posts
 */
export const ListPostsQueryParams = zod.object({
  feed: zod.enum(["foryou", "following", "investors", "hiring"]).optional(),
});

export const ListPostsResponseItem = zod.object({
  id: zod.string(),
  authorId: zod.string(),
  text: zod.string(),
  imageKey: zod.string().nullish(),
  sponsored: zod.boolean(),
  sponsorLabel: zod.string().nullish(),
  category: zod.string(),
  likesCount: zod.number(),
  retweetsCount: zod.number(),
  commentsCount: zod.number(),
  tipsTotal: zod.number(),
  createdAt: zod.coerce.date(),
  liked: zod.boolean(),
  retweeted: zod.boolean(),
});
export const ListPostsResponse = zod.array(ListPostsResponseItem);

/**
 * @summary Create a post
 */
export const CreatePostBody = zod.object({
  text: zod.string(),
  imageKey: zod.string().nullish(),
});

export const ToggleLikeParams = zod.object({
  id: zod.coerce.string(),
});

export const ToggleLikeResponse = zod.object({
  id: zod.string(),
  authorId: zod.string(),
  text: zod.string(),
  imageKey: zod.string().nullish(),
  sponsored: zod.boolean(),
  sponsorLabel: zod.string().nullish(),
  category: zod.string(),
  likesCount: zod.number(),
  retweetsCount: zod.number(),
  commentsCount: zod.number(),
  tipsTotal: zod.number(),
  createdAt: zod.coerce.date(),
  liked: zod.boolean(),
  retweeted: zod.boolean(),
});

export const ToggleRetweetParams = zod.object({
  id: zod.coerce.string(),
});

export const ToggleRetweetResponse = zod.object({
  id: zod.string(),
  authorId: zod.string(),
  text: zod.string(),
  imageKey: zod.string().nullish(),
  sponsored: zod.boolean(),
  sponsorLabel: zod.string().nullish(),
  category: zod.string(),
  likesCount: zod.number(),
  retweetsCount: zod.number(),
  commentsCount: zod.number(),
  tipsTotal: zod.number(),
  createdAt: zod.coerce.date(),
  liked: zod.boolean(),
  retweeted: zod.boolean(),
});

export const TipPostParams = zod.object({
  id: zod.coerce.string(),
});

export const TipPostBody = zod.object({
  amount: zod.number(),
});

export const TipPostResponse = zod.object({
  id: zod.string(),
  authorId: zod.string(),
  text: zod.string(),
  imageKey: zod.string().nullish(),
  sponsored: zod.boolean(),
  sponsorLabel: zod.string().nullish(),
  category: zod.string(),
  likesCount: zod.number(),
  retweetsCount: zod.number(),
  commentsCount: zod.number(),
  tipsTotal: zod.number(),
  createdAt: zod.coerce.date(),
  liked: zod.boolean(),
  retweeted: zod.boolean(),
});

export const ListCirclesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  about: zod.string(),
  category: zod.string(),
  color: zod.string(),
  coverKey: zod.string().nullish(),
  paid: zod.boolean(),
  price: zod.number(),
  membersCount: zod.number(),
  activeNow: zod.number(),
  founderIds: zod.array(zod.string()),
  joined: zod.boolean(),
});
export const ListCirclesResponse = zod.array(ListCirclesResponseItem);

export const ToggleCircleMembershipParams = zod.object({
  id: zod.coerce.string(),
});

export const ToggleCircleMembershipResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  about: zod.string(),
  category: zod.string(),
  color: zod.string(),
  coverKey: zod.string().nullish(),
  paid: zod.boolean(),
  price: zod.number(),
  membersCount: zod.number(),
  activeNow: zod.number(),
  founderIds: zod.array(zod.string()),
  joined: zod.boolean(),
});

export const ListPitchesResponseItem = zod.object({
  id: zod.string(),
  founderId: zod.string(),
  title: zod.string(),
  stage: zod.string(),
  industry: zod.string(),
  raising: zod.number(),
  raised: zod.number(),
  city: zod.string(),
  summary: zod.string(),
  coverKey: zod.string().nullish(),
  backersCount: zod.number(),
  trending: zod.boolean(),
  backed: zod.boolean(),
});
export const ListPitchesResponse = zod.array(ListPitchesResponseItem);

export const BackPitchParams = zod.object({
  id: zod.coerce.string(),
});

export const BackPitchBody = zod.object({
  amount: zod.number(),
});

export const BackPitchResponse = zod.object({
  id: zod.string(),
  founderId: zod.string(),
  title: zod.string(),
  stage: zod.string(),
  industry: zod.string(),
  raising: zod.number(),
  raised: zod.number(),
  city: zod.string(),
  summary: zod.string(),
  coverKey: zod.string().nullish(),
  backersCount: zod.number(),
  trending: zod.boolean(),
  backed: zod.boolean(),
});

export const ListMarkersResponseItem = zod.object({
  id: zod.string(),
  type: zod.string(),
  label: zod.string(),
  city: zod.string(),
  x: zod.number(),
  y: zod.number(),
  meta: zod.string(),
  refId: zod.string().nullish(),
});
export const ListMarkersResponse = zod.array(ListMarkersResponseItem);

export const ListNotificationsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  type: zod.string(),
  actorId: zod.string().nullish(),
  postId: zod.string().nullish(),
  circleId: zod.string().nullish(),
  pitchId: zod.string().nullish(),
  amount: zod.number().nullish(),
  message: zod.string(),
  read: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListNotificationsResponse = zod.array(
  ListNotificationsResponseItem,
);

export const MarkAllNotificationsReadResponse = zod.object({
  ok: zod.boolean(),
});