# lib/api-zod/src/generated/types/circle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HumanVerse API
 * OpenAPI spec version: 0.1.0
 */

export interface Circle {
  id: string;
  name: string;
  about: string;
  category: string;
  color: string;
  /** @nullable */
  coverKey?: string | null;
  paid: boolean;
  price: number;
  membersCount: number;
  activeNow: number;
  founderIds: string[];
  joined: boolean;
}